Refactor all database query code into functions

+++ demo/37-database/functions.php
@@ -1,10 +1,34 @@
 <?php include 'db.php'; ?>
 <?php
 
+function create_rows() {
+	if (isset($_POST['submit'])) {
+		global $connection;
+		$username = $_POST['username'];
+		$password = $_POST['password'];
+
+		$query = 'INSERT INTO users(username, password) ';
+		$query .= 'VALUES (\'' . $username . '\', \'' . $password . '\')';
+
+		$result = mysqli_query($connection, $query);
+
+		if (!$result) {
+			die(
+				'<p class="text-danger mt-3">Query FAILED!</p>' .
+				'<br>' .
+				mysqli_error()
+			);
+		} else {
+			echo '<p class="text-success mt-3">Record created.</p>';
+		}
+	}
+}
+
 function show_all_data() {
 	global $connection;
 
 	$query = 'SELECT * FROM users';
+
 	$result = mysqli_query($connection, $query);
 
 	if (!$result) {
@@ -18,37 +42,53 @@ function show_all_data() {
 }
 
 function update_table() {
-	global $connection;
+	if (isset($_POST['submit'])) {
+		global $connection;
 
-	$username = $_POST['username'];
-	$password = $_POST['password'];
-	$id = $_POST['id'];
+		$username = $_POST['username'];
+		$password = $_POST['password'];
+		$id = $_POST['id'];
 
-	$query = 'UPDATE users SET ';
-	$query .= 'username = \'' . $username . '\', ';
-	$query .= 'password = \'' . $password . '\' ';
-	$query .= 'WHERE id = ' . $id;
+		$query = 'UPDATE users SET ';
+		$query .= 'username = \'' . $username . '\', ';
+		$query .= 'password = \'' . $password . '\' ';
+		$query .= 'WHERE id = ' . $id;
 
-	$result = mysqli_query($connection, $query);
+		$result = mysqli_query($connection, $query);
 
-	if (!$result) {
-		die('QUERY FAILED: ' . mysqli_error($connection));
+		if (!$result) {
+			die(
+				'<p class="text-danger mt-3">Query FAILED!</p>' .
+				'<br>' .
+				mysqli_error()
+			);
+		} else {
+			echo '<p class="text-success mt-3">Record updated.</p>';
+		}
 	}
 }
 
 function delete_rows() {
-	global $connection;
+	if (isset($_POST['submit'])) {
+		global $connection;
 
-	$username = $_POST['username'];
-	$password = $_POST['password'];
-	$id = $_POST['id'];
+		$username = $_POST['username'];
+		$password = $_POST['password'];
+		$id = $_POST['id'];
 
-	$query = 'DELETE FROM users ';
-	$query .= 'WHERE id = ' . $id;
+		$query = 'DELETE FROM users ';
+		$query .= 'WHERE id = ' . $id;
 
-	$result = mysqli_query($connection, $query);
+		$result = mysqli_query($connection, $query);
 
-	if (!$result) {
-		die('QUERY FAILED: ' . mysqli_error($connection));
+		if (!$result) {
+			die(
+				'<p class="text-danger mt-3">Query FAILED!</p>' .
+				'<br>' .
+				mysqli_error()
+			);
+		} else {
+			echo '<p class="text-success mt-3">Record deleted.</p>';
+		}
 	}
 }
